Fixed up Justin's files to make it work again because I changed file paths
