--- client/src/components/ThreadPost.js
@@ -1,18 +1,0 @@
-import React from 'react';
-import PostUserInfo from './PostUserInfo';
-import PostInfo from './PostInfo';
-import './Thread.css';
-
-function ThreadPost(props) {
-  return (
-    <div>
-      <p className="post-date">Posted on: 02-13-2021 9:20am</p>
-      <article className="each-post">
-        <PostUserInfo />
-        <PostInfo />
-      </article>
-    </div>
-  );
-}
-
-export default ThreadPost;
